test: cover missing sounddevice and separate webapp
