+++ tests/test_audio_missing_sounddevice.py
@@ -1,0 +1,9 @@
+import pytest
+
+from OcchioOnniveggente.src import audio
+
+
+def test_record_wav_requires_sounddevice(tmp_path, monkeypatch):
+    monkeypatch.setattr(audio, "sd", None)
+    with pytest.raises(RuntimeError):
+        audio.record_wav(tmp_path / "out.wav", seconds=1, sr=16000)
